+++ Halmap/Data/AudioManager.swift
@@ -1,0 +1,44 @@
+//
+//  AudioManager.swift
+//  Halmap
+//
+//  Created by Yeni Hwang on 2023/03/21.
+//
+
+import AVKit
+import Foundation
+
+final class AudioManager: ObservableObject {
+    static let instance = AudioManager()
+    var player: AVPlayer?
+    
+    @Published private(set) var isPlaying: Bool = false {
+        didSet{
+            print("isPlaying", isPlaying)
+        }
+    }
+    
+    func playSound(_ urlString : String?) {
+        
+        guard let urlString = urlString else { fatalError("url을 받아올 수 없습니다.") }
+        guard let url = URL(string: urlString) else { fatalError("url을 변환할 수 없습니다.") }
+        let item = AVPlayerItem(url: url)
+        
+        player = AVPlayer(playerItem: item)
+        
+        do {
+            try AVAudioSession.sharedInstance().setCategory(.playback)
+        } catch(let error) {
+            print(error.localizedDescription)
+        }
+        player?.play()
+    }
+    
+    func stopSound() {
+        guard let player = player else {
+            print("Instance of audio player not found")
+            return
+        }
+        player.pause()
+    }
+}
